requirements.txt
Django
sorl-thumbnail
django-markitup
django-model-utils
